Extract the Collection widget and add unit tests
